Update to populating-fields.md.
Updates were made to the `Populating createdBy and updatedBy` section because the steps proposed in that section was incomplete. The default behavior in strapiV4 seems to have changed and hence following the proposed steps didn't really do anything. There were additional steps required which I found was mentioned in an issue thread but has no mention of it in the documentation. Hence, I decided to add the steps myself and make the push request to make it easier for anyone else newly coming into this package to wring out the best that Strapi is able to deliver without any frustration.